tests/frontend/test_app.py
import pytest
from unittest.mock import patch, Mock
from frontend.app import make_request

@patch('frontend.app.requests.request')
def test_make_request_successful_response(mock_request):
    # Arrange
    mock_response = Mock()
    mock_response.json.return_value = {'data': 'value'}
    mock_response.raise_for_status = Mock()
    mock_request.return_value = mock_response
    
    # Act
    result = make_request('GET', '/test-endpoint')
    
    # Assert
    assert result == {'data': 'value'}
    mock_request.assert_called_once_with('GET', 'http://BACKEND_URL/test-endpoint', **{})

import pytest
from unittest.mock import patch
from frontend.app import make_request

@patch('frontend.app.requests.request')
def test_make_request_request_exception(mock_request, monkeypatch):
    # Arrange
    mock_request.side_effect = Exception('Network error')
    monkeypatch.setattr('streamlit.error', lambda x: None)  # Mock st.error to prevent output during tests
    
    # Act
    result = make_request('GET', '/test-endpoint')
    
    # Assert
    assert result is None
    mock_request.assert_called_once_with('GET', 'http://BACKEND_URL/test-endpoint', **{})

import pytest
from frontend.app import UserManager
from unittest.mock import patch, MagicMock

@patch('frontend.app.make_request')
def test_UserManager_fetch_and_store_users_scenario1(mock_make_request):
    mock_make_request.return_value = {"message": "Users fetched successfully!", "run_id": "12345"}
    user_manager = UserManager()
    user_manager.fetch_and_store_users(10)
    mock_make_request.assert_called_once_with("POST", "/users/", json={"num_users": 10})
    assert mock_make_request.return_value["message"] == "Users fetched successfully!"
    assert mock_make_request.return_value["run_id"] == "12345"

import pytest
from frontend.app import UserManager
from unittest.mock import patch, MagicMock

@patch('frontend.app.make_request')
def test_UserManager_fetch_and_store_users_scenario2(mock_make_request):
    mock_make_request.return_value = None
    user_manager = UserManager()
    user_manager.fetch_and_store_users(5)
    mock_make_request.assert_called_once_with("POST", "/users/", json={"num_users": 5})
    # Although result is None, we need to assert that there are no exceptions raised
    assert True

import pytest
from frontend.app import UserManager
from unittest.mock import patch

def test_UserManager_get_random_user_success():
    mock_user = {
        'first_name': 'John',
        'last_name': 'Doe',
        'email': 'johndoe@example.com',
        'gender': 'Male',
        'uid': '12345',
        'latitude': 37.7749,
        'longitude': -122.4194,
        'datetime': '2021-01-01T00:00:00'
    }
    with patch('app.make_request', return_value=mock_user):
        user_manager = UserManager()
        user_manager.get_random_user()  # Invoke the method
        # Assertions for Streamlit writes (Mock st.write)
        with patch('streamlit.write') as mock_write:
            user_manager.get_random_user()
            mock_write.assert_any_call('Name: John Doe')
            mock_write.assert_any_call('Email: johndoe@example.com')
            mock_write.assert_any_call('Gender: Male')
            mock_write.assert_any_call('UID: 12345')
            mock_write.assert_any_call('Ingestion Date: 2021-01-01T00:00:00')

import pytest
from frontend.app import UserManager
from unittest.mock import patch


def test_UserManager_get_random_user_no_user_found():
    with patch('app.make_request', return_value=None):
        user_manager = UserManager()
        with patch('streamlit.write') as mock_write:
            user_manager.get_random_user()  # Invoke the method
            mock_write.assert_not_called()

import pytest
from frontend.app import UserManager
from unittest.mock import patch

class TestUserManager:
    @patch('frontend.app.make_request')
    def test_get_random_username_success(self, mock_make_request):
        mock_make_request.return_value = 'test_user123'
        user_manager = UserManager()
        user_manager.get_random_username()
        mock_make_request.assert_called_once_with('GET', '/random_username/')
        # Assume st.success is a callable that we can verify
        assert mock_success.call_count == 1
        mock_success.assert_called_with('Random Username: test_user123')


import pytest
from frontend.app import UserManager
from unittest.mock import patch

class TestUserManager:
    @patch('frontend.app.make_request')
    def test_get_random_username_failure(self, mock_make_request):
        mock_make_request.return_value = None
        user_manager = UserManager()
        user_manager.get_random_username()
        mock_make_request.assert_called_once_with('GET', '/random_username/')
        # Check if success was not called in this case
        assert mock_success.call_count == 0


import pytest
from frontend.app import UserManager

@pytest.fixture
def mock_make_request(monkeypatch):
    def mock_get(*args, **kwargs):
        return [{'first_name': 'John', 'last_name': 'Doe', 'email': 'john@example.com', 'latitude': 34.0522, 'longitude': -118.2437}]
    monkeypatch.setattr('module_containing_make_request.make_request', mock_get)

def test_UserManager_find_nearest_users_scenario1(mock_make_request):
    user_manager = UserManager()
    user_manager.find_nearest_users('valid@example.com', 10)
    # Assertions here to check for expected behavior, like calling st.write or other changes
    # Example: assert some_condition


import pytest
from frontend.app import UserManager

@pytest.fixture
def mock_make_request(monkeypatch):
    def mock_get(*args, **kwargs):
        return None
    monkeypatch.setattr('module_containing_make_request.make_request', mock_get)

def test_UserManager_find_nearest_users_scenario2(mock_make_request):
    user_manager = UserManager()
    user_manager.find_nearest_users('invalid-email', 10)
    # Assertions to check that an error message was triggered
    # Example: assert some_error_triggered_condition


import pytest
from frontend.app import UserAnalytics

# Mocking the make_request function
def mock_make_request(method, endpoint):
    return 10  # Simulating a response of 10 users

@pytest.mark.parametrize('expected_count', [10])
def test_UserAnalytics_get_user_count_scenario1(monkeypatch, expected_count):
    monkeypatch.setattr('frontend.app.make_request', mock_make_request)
    analytics = UserAnalytics()
    analytics.get_user_count()
    assert 'Total number of users: 10' in capsys.readouterr().out


import pytest
from frontend.app import UserAnalytics

# Mocking the make_request function
def mock_make_request_no_users(method, endpoint):
    return None  # Simulating no users found

@pytest.mark.parametrize('expected_output', [None])
def test_UserAnalytics_get_user_count_scenario2(monkeypatch, expected_output):
    monkeypatch.setattr('frontend.app.make_request', mock_make_request_no_users)
    analytics = UserAnalytics()
    analytics.get_user_count()
    assert 'Total number of users:' not in capsys.readouterr().out


import pytest
from frontend.app import UserAnalytics
from unittest.mock import patch

@patch('frontend.app.make_request')
def test_UserAnalytics_get_gender_distribution_scenario1(mock_make_request):
    mock_make_request.return_value = {'male': 100, 'female': 120}
    user_analytics = UserAnalytics()
    user_analytics.get_gender_distribution()
    mock_make_request.assert_called_once_with('GET', '/gender_distribution/')
    # Here you would check the console output or whatever method you use to capture it
    # For example: assert captured_output == 'Gender Distribution:\nmale: 100\nfemale: 120\n'


import pytest
from frontend.app import UserAnalytics
from unittest.mock import patch

@patch('frontend.app.make_request')
def test_UserAnalytics_get_gender_distribution_scenario2(mock_make_request):
    mock_make_request.return_value = None
    user_analytics = UserAnalytics()
    user_analytics.get_gender_distribution()
    mock_make_request.assert_called_once_with('GET', '/gender_distribution/')
    # Again, verify no output or other behavior when there is no data


